Lifted the C++11 requirement for some functions  that take std::string as an argument.
