+++ source/utf8/cpp17.h
@@ -28,17 +28,10 @@ DEALINGS IN THE SOFTWARE.
 #ifndef UTF8_FOR_CPP_7e906c01_03a3_4daf_b420_ea7ea952b3c9
 #define UTF8_FOR_CPP_7e906c01_03a3_4daf_b420_ea7ea952b3c9
 
-#include "checked.h"
-#include <string>
+#include "cpp11.h"
 
 namespace utf8
 {
-
-    inline void append(char32_t cp, std::string& s)
-    {
-        append(cp, std::back_inserter(s));
-    }
-
     inline std::string utf16to8(std::u16string_view s)
     {
         std::string result;
